requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
jinja2==3.1.4
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.0
respx==0.20.1
responses==0.23.1
pytest-mock==3.12.0
freezegun==1.2.2
qrcode==7.4
Pillow==10.1.0
portalocker==2.8.0
pydantic-settings>=2.0.0
mercadopago==2.3.0
bcrypt==4.2.1
python-multipart==0.0.9
resend==2.28.0
